requirements.txt
pycryptodome
filetype